requirements.txt
Django>=3.2.0,<3.10.0
djangorestframework>=3.11.0,<3.12.0
django-cors-headers>=3.2.1,<3.10.0
psycopg2>=2.8.4,<2.9.0
flake8>=3.7.9,<3.8.0
isort>=5.10.0,<5.20.0
coverage==6.2.0
black>=22.10.0,<210.0.0
pydantic>=1.9.0,<1.10.4
python-environ>=0.4.54,<0.5.0
django-import-export>=2.7.0,<2.8.0
Pillow>=7.1.2,<7.2.0
gunicorn==20.1.0
whitenoise==6.2.0
python-dateutil==2.8.2 
six==1.16.0
django-filter==2.4.0
